k8s: fix secret handling
Azure secrets need namespace with newer versions of ansible (2.10)

Also use 'never' tag with k8s-secrets since this should only be run
once because it updates the k8s cluster(s) directly, not just the k8s
runner node.

By adding 'never' to the 'k8s-secrets' tags, this means that in order
for the 'k8s-secrets' tasks to run, "--tags k8s-secrets" has to be
explicitly passed on the ansible command-line.

Signed-off-by: Kevin Hilman <khilman@baylibre.com>